refactor(bridge): separate TransactionLocatoin type
