+++ bridge/src/types/transaction-location.ts
@@ -1,0 +1,7 @@
+import { BlockHash } from "./block-hash";
+import { TxId } from "./txid";
+
+export interface TransactionLocation {
+    blockHash: BlockHash;
+    txId: TxId;
+}
